pyproject.toml
[tool.poetry]
name = "starlite-jwt-auth"
version = "1.0.0"
description = "A JWT auth toolkit for Starlite"
authors = ["Na'aman Hirschfeld <nhirschfeld@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
starlite = "*"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pylint.MESSAGE_CONTROL]
disable = [
    "line-too-long",
]
enable = "useless-suppression"
extension-pkg-allow-list = ["pydantic"]

[tool.pylint.REPORTS]
reports = "no"

[tool.pylint.FORMAT]
max-line-length = "120"

[tool.pylint.VARIABLES]
ignored-argument-names = "args|kwargs|_|__"

[tool.pylint.BASIC]
good-names = "_,__,i,e"
# Regular expression which should only match function or class names that do
# not require a docstring.
no-docstring-rgx="(__.*__|main|test.*|.*test|.*Test|^_.*)$"



[tool.coverage.run]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'if TYPE_CHECKING:',
    'except ImportError as e:',
    '\.\.\.'
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pycln]
all = true